fixIssues
