+++ package.json
@@ -10,20 +10,14 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@emotion/react": "^11.11.4",
-    "@emotion/styled": "^11.11.5",
     "@fortawesome/fontawesome-free": "^6.5.2",
-    "@mui/base": "^5.0.0-beta.40",
     "@mui/icons-material": "^5.15.15",
     "@mui/material": "^5.15.15",
     "@mui/styled-engine-sc": "^6.0.0-alpha.18",
     "@mui/system": "^5.15.15",
     "@mui/x-charts": "^7.3.0",
-    "@mui/x-date-pickers": "^7.3.1",
-    "@mui/x-date-pickers-pro": "^7.3.1",
     "axios": "^1.6.8",
     "bootstrap": "^5.3.3",
-    "dayjs": "^1.11.10",
     "jsonwebtoken": "^9.0.2",
     "jwt-decode": "^4.0.0",
     "react": "^18.2.0",
